pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "radar-marca"
version = "0.4.0"
description = "Detección de presencia digital, dominios sospechosos y posibles suplantaciones de marca."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "Quitiweb" }
]
dependencies = [
  "Flask>=3.1.0",
  "dnspython>=2.8.0",
]

[project.scripts]
radar-marca = "radar_marca.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
